requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-countries==7.0
django-extensions==3.1.2
django-filter==2.4.0
django-heroku==0.3.1
django-measurement==3.2.3
django-nested-admin==3.3.3
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-redis==4.12.1
django-rest-auth==0.9.5
django-versatileimagefield==2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
libmagic==1.0
Markdown==3.3.3
MarkupSafe==1.1.1
measurement==3.2.0
mpmath==1.2.1
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.18
Pillow==8.1.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
python-magic==0.4.22
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
sympy==1.7.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
Faker